Heart/requirement_package.txt
antlr4-python3-runtime==4.9.3
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
gensim==4.3.2
joblib==1.3.2
numpy==1.26.1
omegaconf==2.3.0
packaging==24.0
pandas==2.2.1
pillow==10.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.10.3
scikit-learn==1.4.1.post1
scipy==1.11.3
six==1.16.0
smart-open==6.4.0
threadpoolctl==3.4.0
tqdm==4.66.1
tzdata==2024.1